+++ cpp_src/core/index/index.h
@@ -23,7 +23,8 @@ class RdxContext;
 class Index {
 public:
 	struct SelectOpts {
-		SelectOpts() : distinct(0), disableIdSetCache(0), forceComparator(0), unbuiltSortOrders(0) {}
+		SelectOpts() : itemsCountInNamespace(0), distinct(0), disableIdSetCache(0), forceComparator(0), unbuiltSortOrders(0) {}
+		unsigned itemsCountInNamespace;
 		unsigned distinct : 1;
 		unsigned disableIdSetCache : 1;
 		unsigned forceComparator : 1;